Fix bug in clang-tidy configuration file  (#4321)
* fix issue in clang-tidy

* address most of the issues found with clang-tidy

* address issues found with clang-tidy

* address issues found with clang-tidy

* address issues found with clang-tidy

* fix issues found with clang-tidy

* fix issues found with clang-tidy

* refactor variable name

* fix residual issues found with clang-tidy

* revert jn -> cyl_bessel change

* cleaning

* fix bug

* fix issues found by clang-tidy in new code

* address issues found by clang-tidy in new code

* fixed issues

* address residual issues found with clang-tidy

+++ Source/Particles/Gather/GetExternalFields.cpp
@@ -13,7 +13,7 @@
 
 using namespace amrex::literals;
 
-GetExternalEBField::GetExternalEBField (const WarpXParIter& a_pti, int a_offset) noexcept
+GetExternalEBField::GetExternalEBField (const WarpXParIter& a_pti, long a_offset) noexcept
 {
     auto& warpx = WarpX::GetInstance();
     auto& mypc = warpx.GetPartContainer();
@@ -22,7 +22,7 @@ GetExternalEBField::GetExternalEBField (const WarpXParIter& a_pti, int a_offset)
 
     AcceleratorLattice const & accelerator_lattice = warpx.get_accelerator_lattice(lev);
     if (accelerator_lattice.m_lattice_defined) {
-        d_lattice_element_finder = accelerator_lattice.GetFinderDeviceInstance(a_pti, a_offset);
+        d_lattice_element_finder = accelerator_lattice.GetFinderDeviceInstance(a_pti, static_cast<int>(a_offset));
     }
 
     m_gamma_boost = WarpX::gamma_boost;
